Moved files to appropriate subfolders.
