+++ include/steered_wheel_base_controller/joint_types/vel_joint.h
@@ -1,5 +1,5 @@
-#ifndef _SWBC_VEL_JOINT_H_
-#define _SWBC_VEL_JOINT_H_
+#ifndef _SWBC_JOINT_TYPES_VEL_JOINT_H_
+#define _SWBC_JOINT_TYPES_VEL_JOINT_H_
 
 #include <boost/shared_ptr.hpp>
 #include <ros/ros.h>
