+++ include/steered_wheel_base_controller/steered_wheel_base_controller.h
@@ -35,9 +35,9 @@
 #include <urdf/model.h>
 
 #include "steered_wheel_base_controller/joint_base.h"
-#include "steered_wheel_base_controller/pos_joint.h"
-#include "steered_wheel_base_controller/vel_joint.h"
-#include "steered_wheel_base_controller/pid_joint.h"
+#include "steered_wheel_base_controller/joint_types/pos_joint.h"
+#include "steered_wheel_base_controller/joint_types/vel_joint.h"
+#include "steered_wheel_base_controller/joint_types/pid_joint.h"
 #include "steered_wheel_base_controller/wheel.h"
 #include "steered_wheel_base_controller/util.h"
