+++ plugins/sw_base_controller_plugin.xml
